refactor: move `EventNamespaceStatus` to `plugin-context.ts`
