+++ core/src/events/events.ts
@@ -13,9 +13,8 @@ import type { DeployStatusForEventPayload } from "../types/service.js"
 import type { RunStatusForEventPayload } from "../plugin/base.js"
 import type { Omit, PickFromUnion } from "../util/util.js"
 import type { ConfigGraph, RenderedActionGraph } from "../graph/config-graph.js"
-import type { CommandInfo } from "../plugin-context.js"
+import type { CommandInfo, EventNamespaceStatus } from "../plugin-context.js";
 import type { GraphResult } from "../graph/results.js"
-import type { EventNamespaceStatus } from "../types/namespace.js"
 import type { BuildStatusForEventPayload } from "../plugin/handlers/Build/get-status.js"
 import type { ActionStatusPayload } from "./action-status-events.js"
 import type { AuthToken } from "../cloud/auth.js"
