+++ core/src/plugin-context.ts
@@ -8,14 +8,18 @@
 
 import type { Garden } from "./garden.js"
 import type { SourceConfig } from "./config/project.js"
-import { projectApiVersionSchema } from "./config/project.js"
-import { projectNameSchema, projectSourcesSchema, environmentNameSchema } from "./config/project.js"
-import type { Provider, BaseProviderConfig } from "./config/provider.js"
+import {
+  environmentNameSchema,
+  projectApiVersionSchema,
+  projectNameSchema,
+  projectSourcesSchema,
+} from "./config/project.js"
+import type { BaseProviderConfig, Provider } from "./config/provider.js"
 import { providerSchema } from "./config/provider.js"
 import { deline } from "./util/string.js"
-import { joi, joiVariables, joiStringMap, joiIdentifier, createSchema } from "./config/common.js"
+import { createSchema, joi, joiIdentifier, joiStringMap, joiVariables } from "./config/common.js"
 import type { PluginTool } from "./util/ext-tools.js"
-import type { ContextWithSchema, ContextResolveOpts } from "./config/template-contexts/base.js"
+import type { ContextResolveOpts, ContextWithSchema } from "./config/template-contexts/base.js"
 import { legacyResolveTemplateString } from "./template/templated-strings.js"
 import type { Log } from "./logger/log-entry.js"
 import { logEntrySchema } from "./plugin/base.js"
@@ -24,9 +28,9 @@ import type { CreateEventLogParams, StringLogLevel } from "./logger/logger.js"
 import { EventLogger, LogLevel } from "./logger/logger.js"
 import { Memoize } from "typescript-memoize"
 import type { ParameterObject, ParameterValues } from "./cli/params.js"
-import type { EventNamespaceStatus } from "./types/namespace.js"
 import type { ParsedTemplate, ResolvedTemplate } from "./template/types.js"
 import { deepEvaluate } from "./template/evaluate.js"
+import { NamespaceStatus } from "./types/namespace.js"
 
 export type WrappedFromGarden = Pick<
   Garden,
@@ -130,6 +134,8 @@ export type PluginEventLogMessage = PluginEventLogContext & {
   /** log message */
   msg: string
 }
+
+export type EventNamespaceStatus = Omit<NamespaceStatus, "namespaceUid">
 
 // Define your emitter's types as follows:
 // Key: Event name; Value: Listener function signature
