+++ core/test/integ/src/plugins/kubernetes/kubernetes-type/handlers.ts
@@ -49,9 +49,9 @@ import {
   GardenApiVersion,
 } from "../../../../../../src/constants.js"
 import type { ActionModeMap } from "../../../../../../src/actions/types.js"
-import type { EventNamespaceStatus } from "../../../../../../src/types/namespace.js"
 import stripAnsi from "strip-ansi"
 import type { DeployActionConfig } from "../../../../../../src/actions/deploy.js"
+import { EventNamespaceStatus } from "../../../../../../src/plugin-context.js";
 
 describe("kubernetes-type handlers", () => {
   let tmpDir: tmp.DirectoryResult
